+++ src/instructions/tables.rs
@@ -1,6 +1,7 @@
 //! Functions to load GDT, IDT, and TSS structures.
 
 use crate::structures::gdt::SegmentSelector;
+use crate::VirtAddr;
 
 pub use crate::structures::DescriptorTablePointer;
 
@@ -18,7 +19,7 @@ pub use crate::structures::DescriptorTablePointer;
 #[inline]
 pub unsafe fn lgdt(gdt: &DescriptorTablePointer) {
     #[cfg(feature = "inline_asm")]
-    asm!("lgdt [{}]", in(reg) gdt, options(nostack));
+    asm!("lgdt [{}]", in(reg) gdt, options(readonly, nostack, preserves_flags));
 
     #[cfg(not(feature = "inline_asm"))]
     crate::asm::x86_64_asm_lgdt(gdt as *const _);
@@ -38,10 +39,27 @@ pub unsafe fn lgdt(gdt: &DescriptorTablePointer) {
 #[inline]
 pub unsafe fn lidt(idt: &DescriptorTablePointer) {
     #[cfg(feature = "inline_asm")]
-    asm!("lidt [{}]", in(reg) idt, options(nostack));
+    asm!("lidt [{}]", in(reg) idt, options(readonly, nostack, preserves_flags));
 
     #[cfg(not(feature = "inline_asm"))]
     crate::asm::x86_64_asm_lidt(idt as *const _);
+}
+
+/// Get the address of the current IDT.
+#[inline]
+pub fn sidt() -> DescriptorTablePointer {
+    let mut idt: DescriptorTablePointer = DescriptorTablePointer {
+        limit: 0,
+        base: VirtAddr::new(0),
+    };
+    unsafe {
+        #[cfg(feature = "inline_asm")]
+        asm!("sidt [{}]", in(reg) &mut idt, options(nostack, preserves_flags));
+
+        #[cfg(not(feature = "inline_asm"))]
+        crate::asm::x86_64_asm_sidt(&mut idt as *mut _);
+    }
+    idt
 }
 
 /// Load the task state register using the `ltr` instruction.
@@ -54,8 +72,8 @@ pub unsafe fn lidt(idt: &DescriptorTablePointer) {
 #[inline]
 pub unsafe fn load_tss(sel: SegmentSelector) {
     #[cfg(feature = "inline_asm")]
-    asm!("ltr {0:x}", in(reg) sel.0, options(nostack, nomem));
+    asm!("ltr {0:x}", in(reg) sel.0, options(nomem, nostack, preserves_flags));
 
     #[cfg(not(feature = "inline_asm"))]
-    crate::asm::x86_64_asm_ltr(sel.0)
+    crate::asm::x86_64_asm_ltr(sel.0);
 }